remove get_utils.c
