+++ src/parser/parse_obj.c
@@ -45,3 +45,23 @@ t_object	parse_plane(char *str, t_errors *ecode)
 			get_vector(args[2], ecode, 0), get_color(args[3], ecode));
 	return (ft_free_stra(args), plane);
 }
+
+t_object	parse(char *str, t_errors *ecode)
+{
+	char	*set;
+
+	set = ft_substr(str, 0, 2);
+	if (!ft_strncmp(set, "sp", 3))
+		return (free(set), parse_sphere(str, ecode));
+	if (!ft_strncmp(set, "cy", 3))
+		return (free(set), parse_cylinder(str, ecode));
+	if (!ft_strncmp(set, "pl", 3))
+		return (free(set), parse_plane(str, ecode));
+	return (free(set), parse_sphere(str, ecode));
+}
+
+int	is_object(char *str)
+{
+	return (!ft_strncmp(str, "A", 1) || !ft_strncmp(str, "C", 1)
+		|| !ft_strncmp(str, "L", 1));
+}
